Przemianowanie byte na integery, otestowanie sprawdzania walki w automacie komórkowym

+++ SKAZA/src/main/java/SKAZA/core/math/constants/MapConstants.java
@@ -2,15 +2,15 @@ package SKAZA.core.math.constants;
 
 public class MapConstants {
 	
-	public static Byte numberOfUpperCornerNeighbours = 2;
-	public static Byte numberOfLowerCornerNeighbours = 3;
-	public static Byte numberOfOuterBaseNeighbours = 3;
-	public static Byte numberOfInnerBaseNeighbours = 5;
-	public static Byte numberOfSideNeighbours = 4;
-	public static Byte numberOfNormalCellNeighbours = 6;
+	public static Integer numberOfUpperCornerNeighbours = 2;
+	public static Integer numberOfLowerCornerNeighbours = 3;
+	public static Integer numberOfOuterBaseNeighbours = 3;
+	public static Integer numberOfInnerBaseNeighbours = 5;
+	public static Integer numberOfSideNeighbours = 4;
+	public static Integer numberOfNormalCellNeighbours = 6;
 	
-	public static Byte gridHeight = 13;	//Tylko nieparzyste
-	public static Byte gridWidth = 19;	//Tylko nieparzyste
+	public static Integer gridHeight = 13;	//Tylko nieparzyste
+	public static Integer gridWidth = 19;	//Tylko nieparzyste
 	
 	
 }
